+++ test/case/translation_placeholder_case/liquid_tab/case/case_liquid_tab_2.dart
@@ -1,0 +1,26 @@
+import '../../../../mock_uuid.dart';
+import '../../case.dart';
+
+class CaseLiquidTab2 implements Case {
+  /// Liquid `{% tab "标题" %}` 语法 2 - 跳过翻译
+  const CaseLiquidTab2();
+
+  @override
+  String testText() {
+    return '''
+{% tab "标题" %}
+
+{% tab "Title" %}
+''';
+  }
+
+  @override
+  String expectText() {
+    return '''
+{% tab "标题" %}
+
+<!-- {% tab "Title" %} -->
+{% tab "${MockUuid.translationChunkId}" %}
+''';
+  }
+}