+++ test/case/translation_placeholder_case/top_metadata/case/case_top_metadata_list_item_1.dart
@@ -1,0 +1,40 @@
+import 'package:cfug_translator_bot/src/services/translation_service/translation_placeholder/translation_placeholder.dart';
+
+import '../../../../mock_uuid.dart';
+import '../../case.dart';
+
+class CaseTopMetadataListItem1 implements Case {
+  /// 顶部元数据带列表项 1
+  const CaseTopMetadataListItem1();
+
+  @override
+  String testText() {
+    return '''
+---
+title: More thoughts about performance
+tag:
+  - tag 1
+  - tag 2
+  - tag 3
+description: What is performance, and why is performance important
+---
+''';
+  }
+
+  @override
+  String expectText() {
+    return '''
+---
+# title: More thoughts about performance
+title: ${MockUuid.translationChunkId}
+tag:
+  - tag 1
+  - tag 2
+  - tag 3
+# description: What is performance, and why is performance important
+description: ${MockUuid.translationChunkId}
+---
+${TranslationPlaceholder(MockUuid()).translationNote}
+''';
+  }
+}